+++ frontend-new/goose/src/store/study_page.js
@@ -131,24 +131,28 @@ export default {
         data: credential,
       }).then((res) => {
         dispatch("selectStudy", getters.selectedStudy.id);
+        alert("스터디 정보가 업데이트 되었습니다.")
       });
     },
 
     deleteStudy({ commit, getters,dispatch }) {
-      axios({
-        url: rest.study.study_remove(getters.selectedStudy.id),
-        method: "delete",
-        headers: getters.authHeader,
-      })
-        .then(() => {
-          dispatch('myStudyList')
-          commit("SET_SELECTED_STUDY", {});
-          router.push({ name: "Home" });
+      const check = confirm("정말로 삭제하시겠습니까?")
+      if (check){
+        axios({
+          url: rest.study.study_remove(getters.selectedStudy.id),
+          method: "delete",
+          headers: getters.authHeader,
         })
-        .catch((err) => {
-          alert("스터디 삭제에 실패했습니다.");
-        });
-      router.push({ name: "Home" });
+          .then(() => {
+            dispatch('myStudyList')
+            commit("SET_SELECTED_STUDY", {});
+            router.push({ name: "Home" });
+          })
+          .catch((err) => {
+            alert("스터디 삭제에 실패했습니다.");
+          });
+        router.push({ name: "Home" });
+      }
     },
     myStudyList({ getters, commit }) {
       axios({
@@ -208,18 +212,21 @@ export default {
 
     // 스터디 가입신청 승락
     joinAgree({ getters, dispatch }, credential) {
-      axios({
-        url: rest.study.study_join_agree(),
-        method: "post",
-        headers: getters.authHeader,
-        data: credential,
-      })
+      const check = confirm(`${credential.name}을 받으시겠습니까?`)
+      if (check){
+        axios({
+          url: rest.study.study_join_agree(),
+          method: "post",
+          headers: getters.authHeader,
+          data: credential,
+        })
         .then((res) => {
           dispatch("joinList");
         })
         .catch((err) => {
           console.log(err);
         });
+      }
     },
 
     // 스터디 가입신청 거절
@@ -238,15 +245,21 @@ export default {
     },
 
     dropOutStudy({ getters }, user_pk) {
-      axios({
-        url: rest.study.study_member_out(),
-        method: "delete",
-        headers: {
-          Authorization: getters.token,
-          study_pk: getters.selectedStudy.id,
-          user_pk: user_pk,
-        },
-      });
+      const check = confirm(`${getters.selectedStudy.title}을(를) 탈퇴 하시겠습니까?`)
+      if (check){
+        axios({
+          url: rest.study.study_member_out(),
+          method: "delete",
+          headers: {
+            Authorization: getters.token,
+            study_pk: getters.selectedStudy.id,
+            user_pk: user_pk,
+          },
+        })
+        .then((res)=>{
+          router.push({ name: "Home" });
+        })
+      }
     },
     compile({ getters, dispatch },code) {
       console.log("compile");
